Remove sortOrder from meta attributes list cuz it breaks stuff. Rename the DataProviders to something easier to understand.
